Removed profile socket middleware Token refresh functionality moved into socket middleware
